chore: handle newly introduced circular dep
