+++ src/lib/assets/yaml/overridesFile.ts
@@ -3,7 +3,7 @@ import { CapabilityExport, ModuleConfig } from "../../types";
 import { dumpYaml } from "@kubernetes/client-node";
 import { clusterRole } from "../rbac";
 import { promises as fs } from "fs";
-import { resolveIgnoreNamespaces } from "../webhooks";
+import { resolveIgnoreNamespaces } from "../ignoredNamespaces";
 export type ChartOverrides = {
   apiPath: string;
   capabilities: CapabilityExport[];
